# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Copycast",
    "description": "Self-hosted podcast mirroring and archiving.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/feeds": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "List Feeds (Mirrors and Inboxes)",
        "operationId": "list_feeds",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FeedKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "title",
                "created_at",
                "updated_at",
                "storage_bytes"
              ],
              "type": "string",
              "default": "title",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-capability": "list_feeds"
      }
    },
    "/api/mirrors": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "List Mirrors (alias of list_feeds with kind=mirror)",
        "operationId": "list_mirrors",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "title",
                "created_at",
                "updated_at",
                "storage_bytes"
              ],
              "type": "string",
              "default": "title",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mirrors"
        ],
        "summary": "Create a Mirror; the Catalog is populated before returning",
        "operationId": "create_mirror",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MirrorCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MirrorRead"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-capability": "create_mirror"
      }
    },
    "/api/inboxes": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "List Inboxes (alias of list_feeds with kind=inbox)",
        "operationId": "list_inboxes",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "title",
                "created_at",
                "updated_at",
                "storage_bytes"
              ],
              "type": "string",
              "default": "title",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "inboxes"
        ],
        "summary": "Create an Inbox",
        "operationId": "create_inbox",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboxCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxRead"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "x-capability": "create_inbox"
      }
    },
    "/api/feeds/{feed_id}": {
      "get": {
        "tags": [
          "feeds"
        ],
        "summary": "One Feed",
        "operationId": "get_feed",
        "parameters": [
          {
            "name": "feed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feed Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MirrorRead"
                    },
                    {
                      "$ref": "#/components/schemas/InboxRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "kind",
                    "mapping": {
                      "mirror": "#/components/schemas/MirrorRead",
                      "inbox": "#/components/schemas/InboxRead"
                    }
                  },
                  "title": "Response Get Feed"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-capability": "get_feed"
      },
      "delete": {
        "tags": [
          "feeds"
        ],
        "summary": "Delete a Feed and its files (the default Inbox is refused)",
        "operationId": "delete_feed",
        "parameters": [
          {
            "name": "feed_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Feed Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-capability": "delete_feed"
      }
    },
...
    "/api/admin/rebuild": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Reconstruct the database from the data directory (worker job under WORKER_LOCK)",
        "operationId": "rebuild",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RebuildRequest",
                "default": {
                  "dry_run": true
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-capability": "rebuild"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackfillMode": {
        "type": "string",
        "enum": [
          "all",
          "latest",
          "selection"
        ],
        "title": "BackfillMode"
      },
      "BackfillPolicy": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/BackfillMode"
          },
          "latest_n": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest N"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "BackfillPolicy"
      },
      "BackfillRequest": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/BackfillMode",
            "default": "all"
          },
          "latest_n": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest N"
          },
          "selection": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096
              },
              {
                "type": "null"
              }
            ],
            "title": "Selection"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BackfillRequest",
        "description": "Backfill policy at creation; ``selection`` implies mode ``selection``."
      },
      "CatalogCounts": {
        "properties": {
          "listed": {
            "type": "integer",
            "title": "Listed",
            "default": 0
          },
          "available": {
            "type": "integer",
            "title": "Available",
            "default": 0
          },
          "delisted": {
            "type": "integer",
            "title": "Delisted",
            "default": 0
          },
          "wanted": {
            "type": "integer",
            "title": "Wanted",
            "default": 0
          },
          "archived": {
            "type": "integer",
            "title": "Archived",
            "default": 0
          },
          "failed": {
            "type": "integer",
            "title": "Failed",
            "default": 0
          }
        },
        "type": "object",
        "title": "CatalogCounts"
      },
...
      "ErrorKind": {
        "type": "string",
        "enum": [
          "transient",
          "permanent",
          "cancelled",
          "storage_full"
        ],
        "title": "ErrorKind"
      },
      "FeedHealth": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/HealthStatus"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "FeedHealth"
      },
      "FeedKind": {
        "type": "string",
        "enum": [
          "mirror",
          "inbox"
        ],
        "title": "FeedKind"
      },
      "FeedList": {
        "properties": {
          "feeds": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/MirrorRead"
                },
                {
                  "$ref": "#/components/schemas/InboxRead"
                }
              ],
              "discriminator": {
                "propertyName": "kind",
                "mapping": {
                  "inbox": "#/components/schemas/InboxRead",
                  "mirror": "#/components/schemas/MirrorRead"
                }
              }
            },
            "type": "array",
            "title": "Feeds"
          }
        },
        "type": "object",
        "title": "FeedList"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InboxCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Name"
          },
          "autoprune_days": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Autoprune Days"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "InboxCreate"
      },
      "InboxRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "artwork_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artwork Url"
          },
          "feed_url": {
            "type": "string",
            "title": "Feed Url"
          },
          "episode_count": {
            "type": "integer",
            "title": "Episode Count",
            "default": 0
          },
          "storage_bytes": {
            "type": "integer",
            "title": "Storage Bytes",
            "default": 0
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "kind": {
            "type": "string",
            "const": "inbox",
            "title": "Kind",
            "default": "inbox"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "autoprune_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Autoprune Days"
          },
          "request_count": {
            "type": "integer",
            "title": "Request Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "feed_url",
          "created_at",
          "name"
        ],
        "title": "InboxRead"
      },
...
      "JobKind": {
        "type": "string",
        "enum": [
          "refresh",
          "archive_item",
          "expand_request",
          "prune",
          "rebuild"
        ],
        "title": "JobKind"
      },
...
      "JobRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "feed_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feed Id"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "trigger": {
            "$ref": "#/components/schemas/JobTrigger"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "progress": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobProgress"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "trigger",
          "status",
          "created_at"
        ],
        "title": "JobRead"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus"
      },
      "JobTrigger": {
        "type": "string",
        "enum": [
          "manual",
          "scheduled",
          "feed_fetch",
          "request",
          "policy",
          "ui",
          "mcp"
        ],
        "title": "JobTrigger"
      },
      "MirrorCreate": {
        "properties": {
          "source_url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "title": "Source Url"
          },
          "candidate_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Token"
          },
          "backfill": {
            "$ref": "#/components/schemas/BackfillRequest"
          },
          "follow": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow",
            "description": "Defaults to true, or false under a selection backfill"
          },
          "engine_options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Engine Options"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_url"
        ],
        "title": "MirrorCreate"
      },
      "MirrorRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "artwork_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artwork Url"
          },
          "feed_url": {
            "type": "string",
            "title": "Feed Url"
          },
          "episode_count": {
            "type": "integer",
            "title": "Episode Count",
            "default": 0
          },
          "storage_bytes": {
            "type": "integer",
            "title": "Storage Bytes",
            "default": 0
          },
          "revision": {
            "type": "integer",
            "title": "Revision",
            "default": 1
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "kind": {
            "type": "string",
            "const": "mirror",
            "title": "Kind",
            "default": "mirror"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "service": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service"
          },
          "source_kind": {
            "$ref": "#/components/schemas/SourceKind"
          },
          "paused": {
            "type": "boolean",
            "title": "Paused",
            "default": false
          },
          "follow": {
            "type": "boolean",
            "title": "Follow",
            "default": true
          },
          "backfill": {
            "$ref": "#/components/schemas/BackfillPolicy"
          },
          "engine_options": {
            "additionalProperties": true,
            "type": "object",
            "title": "Engine Options"
          },
          "last_refresh_attempt_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refresh Attempt At"
          },
          "last_refresh_success_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Refresh Success At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "health": {
            "$ref": "#/components/schemas/FeedHealth"
          },
          "counts": {
            "$ref": "#/components/schemas/CatalogCounts"
          },
          "selection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SelectionSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "feed_url",
          "created_at",
          "source_url",
          "source_kind",
          "backfill",
          "health"
        ],
        "title": "MirrorRead"
      },
...
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem details; ``type`` is ``urn:copycast:problem:<slug>``."
      },
...
      "RebuildRequest": {
        "properties": {
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RebuildRequest"
      },
...
      "SelectionSummary": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Items selected for archiving so far"
          },
          "expression": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expression",
            "description": "The last selection expression"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          }
        },
        "type": "object",
        "required": [
          "count"
        ],
        "title": "SelectionSummary",
        "description": "What a selection-mode Mirror has selected so far."
      },
      "SourceKind": {
        "type": "string",
        "enum": [
          "rss",
          "ytdlp"
        ],
        "title": "SourceKind"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}